fix(geth-types): post-review HIGH/MEDIUM/NIT fixes
HIGH:
- BeaconPivot.GetLowestBlockToFinalize: guard `Head.Number - MaxDepth + 1` —
  wrapped to a near-ulong.MaxValue safe number when head < 64.
- Optimism BatchV1.DecodeTxs: drop (ulong) cast on UInt256 tx values; the
  Transaction.Value / DecodedMaxFeePerGas fields are UInt256 and the cast
  silently truncated any tx > ~18.4 ETH.
- EthereumGasPolicy.CreateAvailableFromIntrinsic: Debug.Assert on the two
  unguarded subtracts (gasLimit - intrinsicRegular - intrinsicState and
  TX_MAX_GAS_LIMIT - intrinsicRegular). Callers validate today; this catches
  future eth_call paths that skip validation.
- EvmInstructions.Storage SSTORE refund: Debug.Assert + SaturatingSub on
  `Refund -= sClearRefunds`. Per EIP-2200/3529 the matching `+= sClearRefunds`
  ran earlier so the invariant holds; the saturating subtract guards against
  an out-of-order wrap silently granting the maximum post-cap refund.
- ReceiptMigration.MigrationPointerTracker.ReportCompleted: guard
  `_nextToConfirm--` at 0 so a fully-migrated chain doesn't wrap the
  tracker into ulong.MaxValue. Reports `MigratedBlockNumber = 0` when
  genesis is reached.

MEDIUM:
- VmState.CommitToParent: `checked` the parentState.Refund += Refund. Gas-
  bounded so unreachable in practice, but the unsigned add no longer caught
  a buggy negative-child propagation that master's signed long would have.
- Eip8037BlockGasInclusionCheck.Validate: early-return RegularDimensionExceeded
  / StateDimensionExceeded when cumulative > limit. The prior SaturatingSub
  silently floored to 0 and the worst-case check could falsely succeed.
- CompactionSchedule.NextFullCompactionAfter: clamp `from + distance` at
  ulong.MaxValue. Unreachable at any realistic chain height; explicit.
- AuRaBlockFinalizationManager.LoadInitialLastFinalizedBlockLevel and
  GetFinalizationLevel: drop the long round-trip and use ulong arithmetic
  with explicit zero/underflow guards. Avoids the 2^63-1 wrap that was only
  theoretical but exposed by the signed→unsigned migration.

NIT:
- BlockTree.GetBlockHashOnMainOrBestDifficultyHash: drop dead `blockNumber < 0`
  check on ulong parameter.
- ReceiptsSyncFeed: drop stale "_barrier is long" comment (now ulong).

+++ src/Nethermind/Nethermind.Optimism/CL/Decoding/BatchV1.cs
@@ -116,7 +116,7 @@ public sealed class BatchV1
                         }
 
                         (UInt256 legacyValue, tx.GasPrice, byte[] legacyData) = DecodeLegacyTransaction(Txs.Data[(int)txIdx].Span);
-                        tx.Value = (ulong)legacyValue;
+                        tx.Value = legacyValue;
                         tx.Data = legacyData;
                         break;
                     }
@@ -124,7 +124,7 @@ public sealed class BatchV1
                     {
                         v = EthereumEcdsaExtensions.CalculateV(chainId, parityBit);
                         (UInt256 accessValue, tx.GasPrice, byte[] accessData, tx.AccessList) = DecodeAccessListTransaction(Txs.Data[(int)txIdx].Span);
-                        tx.Value = (ulong)accessValue;
+                        tx.Value = accessValue;
                         tx.Data = accessData;
                         break;
                     }
@@ -133,8 +133,8 @@ public sealed class BatchV1
                         v = EthereumEcdsaExtensions.CalculateV(chainId, parityBit);
                         (UInt256 eipValue, tx.GasPrice, UInt256 maxFee, byte[] eipData, tx.AccessList) =
                             DecodeEip1559Transaction(Txs.Data[(int)txIdx].Span);
-                        tx.Value = (ulong)eipValue;
-                        tx.DecodedMaxFeePerGas = (ulong)maxFee;
+                        tx.Value = eipValue;
+                        tx.DecodedMaxFeePerGas = maxFee;
                         tx.Data = eipData;
                         break;
                     }
